LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/dir-auto-form-associated.window-expected.txt


FAIL <input dir=auto type=hidden> directionality assert_true: expected true got false
FAIL <input dir=auto type=text> directionality assert_true: expected true got false
FAIL <input dir=auto type=search> directionality assert_true: expected true got false
FAIL <input dir=auto type=tel> directionality assert_true: expected true got false
FAIL <input dir=auto type=url> directionality assert_true: expected true got false
FAIL <input dir=auto type=email> directionality assert_true: expected true got false
FAIL <input dir=auto type=password> directionality assert_true: expected true got false
FAIL <input dir=auto type=submit> directionality assert_true: expected true got false
FAIL <input dir=auto type=reset> directionality assert_true: expected true got false
FAIL <input dir=auto type=button> directionality assert_true: expected true got false
FAIL <input dir=auto type=date> directionality assert_equals: expected "date" but got "text"
FAIL <input dir=auto type=month> directionality assert_equals: expected "month" but got "text"
FAIL <input dir=auto type=week> directionality assert_equals: expected "week" but got "text"
FAIL <input dir=auto type=time> directionality assert_equals: expected "time" but got "text"
FAIL <input dir=auto type=datetime-local> directionality assert_equals: expected "datetime-local" but got "text"
PASS <input dir=auto type=number> directionality
PASS <input dir=auto type=range> directionality
FAIL <input dir=auto type=color> directionality assert_equals: expected "color" but got "text"
PASS <input dir=auto type=checkbox> directionality
PASS <input dir=auto type=radio> directionality
PASS <input dir=auto type=image> directionality
PASS <textarea dir=auto> directionality
